Refining connecting examples by putting in same class, renaming the maven project, and reformating the code
