+++ java/connect/README.md
@@ -1,0 +1,12 @@
+# Connecting to Cloud Bigtable
+
+This is a collection of samples showing how to use the Java HBase client to connect to Cloud 
+Bigtable. For more information see [Connecting to Cloud Bigtable](https://cloud.google.com/bigtable/docs/hbase-connecting).
+
+
+## Testing samples 
+
+To run tests:
+```
+mvn test -Dbigtable.projectID=test-project -Dbigtable.instanceID=test-instance
+```
