+++ java/connect/pom.xml
@@ -1,18 +1,39 @@
 <?xml version="1.0"?>
+<!--
+Copyright 2018 Google LLC
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+-->
 <project
   xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd"
   xmlns="http://maven.apache.org/POM/4.0.0"
   xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance">
   <modelVersion>4.0.0</modelVersion>
+
+  <!--
+    The parent pom defines common style checks and testing strategies for code samples.
+    Removing or replacing it should not affect the execution of the samples in anyway.
+  -->
   <parent>
-    <groupId>com.google.cloud</groupId>
-    <artifactId>bigtable-samples</artifactId>
-    <version>1.0.0</version>
+    <groupId>com.google.cloud.samples</groupId>
+    <artifactId>shared-configuration</artifactId>
+    <version>1.0.10</version>
   </parent>
-  <groupId>com.example.cloud.bigtable.connecting</groupId>
-  <artifactId>connecting-to-bigtable</artifactId>
+
+  <groupId>com.example.cloud.bigtable.connect</groupId>
+  <artifactId>connect</artifactId>
   <version>1.0-SNAPSHOT</version>
-  <name>connecting-to-bigtable</name>
+  <name>connect</name>
   <url>http://maven.apache.org</url>
 
   <properties>
@@ -21,7 +42,7 @@
     <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
   </properties>
 
-  <!--START bigtable_connecting_include_resources-->
+  <!--[START bigtable_connect_include_resources]-->
   <build>
     <resources>
       <resource>
@@ -30,7 +51,7 @@
       </resource>
     </resources>
   </build>
-  <!--END bigtable_connecting_include_resources-->
+  <!--END [bigtable_connect_include_resources]-->
 
   <dependencies>
     <dependency>
@@ -41,19 +62,13 @@
     <dependency>
       <groupId>junit</groupId>
       <artifactId>junit</artifactId>
-      <version>3.8.1</version>
-      <scope>test</scope>
-    </dependency>
-    <dependency>
-      <groupId>junit</groupId>
-      <artifactId>junit</artifactId>
       <version>4.12</version>
       <scope>test</scope>
     </dependency>
     <dependency>
       <groupId>com.google.truth</groupId>
       <artifactId>truth</artifactId>
-      <version>RELEASE</version>
+      <version>0.42</version>
       <scope>test</scope>
     </dependency>
   </dependencies>
