+++ java/pom.xml
@@ -13,7 +13,8 @@ distributed under the License is distributed on an "AS IS" BASIS,
 WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
---><project>
+-->
+<project>
   <modelVersion>4.0.0</modelVersion>
   <version>1.0.0</version>
 
@@ -26,10 +27,10 @@ limitations under the License.
     Removing or replacing it should not affect the execution of the samples in anyway.
   -->
   <parent>
-		<groupId>com.google.cloud.samples</groupId>
-		<artifactId>shared-configuration</artifactId>
-		<version>1.0.10</version>
-	</parent>
+    <groupId>com.google.cloud.samples</groupId>
+    <artifactId>shared-configuration</artifactId>
+    <version>1.0.10</version>
+  </parent>
 
   <modules>
     <module>dataflow-coinbase</module>
@@ -44,6 +45,6 @@ limitations under the License.
     <module>simple-performance-test</module>
     <module>dataflow-bulk-delete</module>
     <module>quickstart</module>
-    <module>connecting-to-bigtable</module>
+    <module>connect</module>
   </modules>
 </project>
